backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy>=2.0
pydantic>=2.11
pydantic-settings>=2.6
python-multipart>=0.0.9
alembic>=1.13
bcrypt>=4.1
python-jose[cryptography]>=3.3
httpx>=0.27
lxml>=5.2
cryptography>=42.0
signxml>=3.2
psycopg[binary]>=3.1
jinja2>=3.1
qrcode[pil]>=7.4
google-auth>=2.27
requests>=2.31


